Add all instructions to the decoder

+++ src/il/co/codeguru/corewars8086/cpu/riscv/instruction_formats/InstructionI.java
@@ -37,8 +37,8 @@ public class InstructionI extends InstructionBase {
         return (byte)((this.raw >> 15) & mask(5));
     }
 
-    public byte getImmediate()
+    public short getImmediate()
     {
-        return (byte)(this.raw >> 20);
+        return (short)(this.raw >> 20);
     }
 }
